Draw sprites on scanlines above the vanilla view
Sprites vanished above the original 240x160 area but drew correctly below it.
DrawAffineSprite and DrawNonAffineSprite declared uint16_t vcount, so a row
above the view -- scanline -12 -- arrived as 65524, and the sprite's row test
(vcount >= y - half_height && vcount < y + half_height) compared that against
small negatives and never matched. No sprite could draw on a negative scanline
whatever the window state; rows below were fine because their vcount is
positive. DrawSprites and DrawSpritesWinMask were widened to int when sprite
clipping was first fixed, and these two callees were missed.

Also drops the OAM wrap convention from SortSprites. oam.y is a real signed
coordinate now, so subtracting 256 from anything at y >= 160 sorted sprites
below the view as if they were above the screen, ordering them behind sprites
and map layers they should draw in front of. Both copies in the insertion sort
and the double-size affine variants are gone.

+++ src/platform/gba_fast_draw.c
@@ -2063,7 +2063,7 @@ static void DrawSpritesWinMask(struct scanlineData* scanline, int vcount)
     }
 }
 
-static void inline_hack DrawAffineSprite(int SpriteIndex, struct scanlineData* scanline, uint16_t vcount, bool windowsEnabled, uint16_t* pixels, bool IsInsideWinIn)
+static void inline_hack DrawAffineSprite(int SpriteIndex, struct scanlineData* scanline, int vcount, bool windowsEnabled, uint16_t* pixels, bool IsInsideWinIn)
 {
     struct OamData *oam = &((struct OamData *)OAM)[SpriteIndex];
     void *objtiles = VRAM_ + 0x10000;
@@ -2282,7 +2282,7 @@ static void inline_hack DrawAffineSprite(int SpriteIndex, struct scanlineData* s
     }
 }
 
-static void inline_hack DrawNonAffineSprite(int SpriteIndex, struct scanlineData* scanline, uint16_t vcount, bool windowsEnabled, uint16_t* pixels, bool IsInsideWinIn)
+static void inline_hack DrawNonAffineSprite(int SpriteIndex, struct scanlineData* scanline, int vcount, bool windowsEnabled, uint16_t* pixels, bool IsInsideWinIn)
 {
     struct OamData *oam = &((struct OamData *)OAM)[SpriteIndex];
     void *objtiles = VRAM_ + 0x10000;
@@ -2893,6 +2893,7 @@ void DrawFrame(uint16_t *pixels)
 {
     int i;
     int j;
+
 
     
     //memsetu16(pixels, *(uint16_t *)PLTT, DISPLAY_WIDTH * DISPLAY_HEIGHT);
